refactor: 🚚 organized the `/components` folder
Related: #179
