+++ src/components/Main/AboutUs/index.tsx
@@ -1,7 +1,6 @@
-'use client';
 import Image from 'next/image';
 
-import Mission from '@/public/Main/mission.svg'
+import Mission from '@/public/Main/mission.svg';
 import Vision from '@/public/Main/vision.svg';
 
 const content = [
