+++ src/components/TeamCard/TeamCard.css
